src/dataNode/requirements.txt
grpcio==1.60.1
grpcio-tools==1.60.1
python-dotenv==1.0.1
requests==2.31.0
fastapi==0.110.0
uvicorn==0.29.0